requirements.txt
azure-functions==1.13.0  # Azure Functions SDK
azure-eventhub==5.11.0   # Azure Event Hub SDK